Update: Eliminate use of package 'periodic';
Eliminated the use of the package 'periodic', which relied on
SQLAlchemy. Some problems using SQLAlchemy in Windows environments was
experienced, and the use of periodic was easily replaced by the
inclusion of an elemental mass dictionary in __init__.py. Imported as
'from pygaero import table_of_elements'.

- Updated _dchck.py: eliminated use of periodic. Now uses
table_of_elements.
- Updated gen_chem.py: eliminated use of periodic.
- Updated __init__.py: defined dictionary 'table_of_elements' that
contains elemental symbols as keys and elemental masses as values.
- Added pygaero/elements.csv: csv file with elemental symbols and
corresponding atomic weights.